fix: update scorer to April 2026 KDD rules + remove dead kdd_mode
scorer.py:
- Formula: binary → Recall - λ·(extra/predicted), λ=1.0 default
- String: remove .lower(), comparison is now case-sensitive per spec
- Numeric: Decimal ROUND_HALF_UP to 2dp (was float with tolerance)
- Null: handle 'null'/'none'/'nan'/'nat'/'<na>' string variants → ''
- Date: normalize to ISO 8601 YYYY-MM-DD
- Extra columns now penalized (not free as previously assumed)

run_eval.py: overall/per-diff now mean float score; failure threshold < 1.0
compare.py: improved/regressed uses float delta > 0.001
types.py: TaskScore.score int → float

synthesizer/kdd_mode.py: delete — expand_kdd_columns() was dead code
built on the wrong 'no penalty for extra columns' assumption

CLAUDE.md: update scoring description to reflect actual formula

+++ dataline/core/types.py
@@ -115,7 +115,7 @@ class LLMUsage:
 class TaskScore:
     """Eval result for a single task."""
     task_id: str
-    score: int  # 0 or 1 for KDD
+    score: float  # 0.0–1.0: Recall − λ·(extra/predicted)
     difficulty: str = ""
     failure_category: str = ""  # code_error | wrong_direction | format_error | ...
     failed_at_agent: str = ""   # which agent failed
